feature/multi-set-enhancements (#2)
* feat(support/multi_set): add delete method

* feat(support/multi_set): Add pop function to remove and return an item

The changes add a new `pop` function to the `multi_set` module. This function
allows removing and returning an item from the multi-set for a given key. If
the key has multiple items, it will return the last item added and update the
multi-set accordingly. If the key has no items, it will return an error.

This feature is useful for cases where you need to remove and retrieve an item
from a multi-set in a single operation.

* feat(support/multi_set): added reverse method

This reverses all lists in the multi_set. By default a multi_set is first-in-last-out. By calling reverse, it turns into a first-in-first-out. Via the pop method, that is.