feat(day_9): Optimize space finding algorithm
Refactors the space finding algorithm to use a more efficient
approach. Instead of creating a list of the available spaces
and checking if they are all empty, the new implementation
checks if the range of spaces is large enough in a single
pass. This reduces the memory usage and improves the overall
performance of the solution.